# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A comprehensive file server API built with Go. It supports file and folder management, sharing, real-time updates via WebSockets, and more. All protected endpoints require a Bearer Token for authorization. The API implements rate limiting: most routes are limited to 60 requests/second per IP, with stricter limits on login endpoints.",
        "title": "File Server API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/nodes/upload/initiate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Starts a new upload session for a large file. Returns an upload_id to be used for subsequent chunk uploads.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "uploads"
                ],
                "summary": "Initiate a chunked file upload",
                "parameters": [
                    {
                        "description": "File metadata",
                        "name": "uploadRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.InitiateUploadRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_api.InitiateUploadResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid name or size",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden - Write permission denied",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found - Parent folder not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict - a file with the same name already exists",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "413": {
                        "description": "Payload Too Large - Storage quota exceeded",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/nodes/upload/{uploadId}": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Uploads a single chunk of a file for a given upload_id. The 'Content-Range' header is required.",
                "consumes": [
                    "application/octet-stream"
                ],
                "tags": [
                    "uploads"
                ],
                "summary": "Upload a file chunk",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The ID of the upload session",
                        "name": "uploadId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Indicates the byte range of the chunk (e.g., 'bytes 0-1048575/4194304')",
                        "name": "Content-Range",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request - Invalid Content-Range header",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found - Upload session not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "416": {
                        "description": "Range Not Satisfiable - The chunk's start byte does not match the expected offset",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/nodes/upload/{uploadId}/complete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Finalizes a chunked upload after all chunks have been sent. The server verifies the file and moves it to permanent storage.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "uploads"
                ],
                "summary": "Complete a chunked upload",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The ID of the upload session",
                        "name": "uploadId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/serwer-plikow_internal_models.RichNode"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Upload incomplete or file mismatch",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found - Upload session not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_api.InitiateUploadRequest": {
            "type": "object",
            "properties": {
                "mime_type": {
                    "type": "string",
                    "example": "application/zip"
                },
                "name": {
                    "type": "string",
                    "example": "annual_report.zip"
                },
                "parent_id": {
                    "type": "string",
                    "example": "_vx2a-43VqRT5wz_s9u4"
                },
                "size": {
                    "type": "integer",
                    "example": 1073741824
                }
            }
        },
        "internal_api.InitiateUploadResponse": {
            "type": "object",
            "properties": {
                "upload_id": {
                    "type": "string",
                    "example": "a1b2c3d4-e5f6-7890-1234-567890abcdef"
                }
            }
        },
...
        "serwer-plikow_internal_models.RichNode": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "_vx2a-43VqRT5wz_s9u4"
                },
                "is_favorited": {
                    "type": "boolean",
                    "example": true
                },
                "is_shared": {
                    "type": "boolean",
                    "example": false
                },
                "mime_type": {
                    "type": "string",
                    "example": "application/pdf"
                },
                "modified_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "example": "Raport_Q3.docx"
                },
                "node_type": {
                    "type": "string",
                    "example": "file"
                },
                "original_parent_id": {
                    "type": "string"
                },
                "owner": {
                    "$ref": "#/definitions/serwer-plikow_internal_models.RichNodeOwner"
                },
                "parent_id": {
                    "type": "string",
                    "example": "fLW5kAh2ia9vYmjMnU4nZ"
                },
                "permissions": {
                    "type": "string"
                },
                "size_bytes": {
                    "type": "integer",
                    "example": 123456
                }
            }
        },
        "serwer-plikow_internal_models.RichNodeOwner": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string",
                    "example": "Administrator"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "username": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}